third_party/WebKit/LayoutTests/fast/events/onunload-expected.txt

main frame - has 1 onunload handler(s)
CONSOLE ERROR: line 16: Blocked alert('unload') during unload.
You should have seen an unload alert appear.
